backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
stripe==12.3.0
python-dotenv==1.0.1
pydantic==2.10.6
